Get Location information directly from the database and other changes
The location for holdings in voyager is an ID and not the code.
Get a list of Codes to ID from voyager first.
Then get the list of codes to UUIDs from folio, mapping the ID from voyager to the UUID from folio.

A locationSql is now needed on the query.
To keep it simple, use `id` and `code` as the respective select columns, so be sure to include `AS id` and `AS code` in the query.

Use `holdings` instead of `instances` for holdings hrid.

I accidentally had both permanentLocationId and locationId on the holdings model.
They are the same, so remove permanentLocationId and leave locationId.