Improvements, Fixes, and tweaks to vendor migrations
Rename `id` to `addressId` where id is an address id.

Move `vendorId` assignment outside of the class instantiation method for classes where it is not part of the intended model.
In these cases, VendorId is used for logging or other miscellaneous purposes.

Closer match the base scripts, do no report errors for unknown cases that are to be ignored.

Reword the E-mail/URL used as a Phone Number error message.

Remove `vendorId` from classes that don't use or need it, even for logging or miscellaneous purposes.